+++ BE/src/main/java/com/enolj/airbnb/web/HouseController.java
@@ -35,46 +35,4 @@ public class HouseController {
         logger.info("{}번 숙소의 디테일 페이지 요청", houseId);
         return houseService.getHouseDetailInfo(houseId);
     }
-
-    @GetMapping("/houses/{houseId}")
-    public ReservationInfoResponseDTO getReservationPage(@PathVariable Long houseId) {
-        logger.info("{}번 숙소 예약 페이지 요청", houseId);
-        return houseService.getReservationInfo(houseId);
-    }
-
-    @PostMapping("houses/{houseId}")
-    public void makeReservation(@RequestHeader String authorization, @PathVariable Long houseId, @RequestBody ReservationRequestDTO requestDTO) {
-        logger.info("{}번 숙소 예약 요청", houseId);
-        houseService.makeReservation(authorization, houseId, requestDTO);
-    }
-
-    @GetMapping("/wishes")
-    public List<WishResponseDTO> getWishes(@RequestHeader String authorization) {
-        logger.info("숙소 위시 리스트 요청");
-        return houseService.getWishList(authorization);
-    }
-
-    @PostMapping("/wishes/{houseId}")
-    public void changeWish(@RequestHeader String authorization, @PathVariable Long houseId) {
-        logger.info("{}번 숙소의 위시 요청", houseId);
-        houseService.changeWish(authorization, houseId);
-    }
-
-    @GetMapping("/reservation")
-    public List<ReservationResponseDTO> getReservationList(@RequestHeader String authorization) {
-        logger.info("숙소 예약 리스트 요청");
-        return houseService.getReservationList(authorization);
-    }
-
-    @GetMapping("/reservation/{houseId}")
-    public ReservationDetailDTO getReservationDetail(@RequestHeader String authorization, @PathVariable Long houseId) {
-        logger.info("{}번 숙소의 디테일 예약정보 요청", houseId);
-        return houseService.getReservationDetail(authorization, houseId);
-    }
-
-    @DeleteMapping("/reservation/{houseId}")
-    public void cancelReservation(@RequestHeader String authorization, @PathVariable Long houseId) {
-        logger.info("{}번 숙소의 예약 취소 요청", houseId);
-        houseService.cancelReservation(authorization, houseId);
-    }
 }
